Extract vector serialization to src/core/vector/serialization.rs
